backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
websockets
google-generativeai
python-dotenv
aiohttp
asyncio
packaging
mediapipe
opencv-python-headless
numpy
ultralytics 
